requirements.txt
torch~=1.13.0+cu116
torchvision~=0.14.0+cu116
numpy~=1.21.6
en-core-web-sm==2.3.0
spacy~=3.4.3
pillow~=9.2.0
matplotlib==3.1.3
transformers~=4.24.0
opencv-python~=4.6.0.66
sklearn

yacs~=0.1.8
timm~=0.6.11
argparse~=1.4.0
tabulate~=0.9.0
ipython~=7.34.0
tqdm~=4.64.0
termcolor~=2.1.0
datasets~=2.5.2
reFILE~=0.4.1
torchmetrics~=0.9.3
scipy~=1.7.3
colorlog~=6.7.0